[IMP] purchase_invoice_plan: split purchase to files with one class only
