+++ purchase_invoice_plan/models/__init__.py
@@ -1,4 +1,5 @@
 # Copyright 2019 Ecosoft Co., Ltd (http://ecosoft.co.th/)
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html)
 
-from . import purchase
+from . import purchase_order
+from . import purchase_invoice_plan
